Finished Noetic compatibility patch series.
mtdevice.py got split in two pieces - a library (in package) and an executable (in nodes/)
