+++ CMakeLists.txt
@@ -3,11 +3,12 @@ project(xsens_driver)
 
 find_package(catkin REQUIRED)
 
+catkin_python_setup()
+
 catkin_package()
 
 catkin_install_python(PROGRAMS
   nodes/mtnode.py
-  nodes/mtdef.py
   nodes/mtdevice.py
   DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
